+++ examples/squelch.ml
@@ -7,7 +7,7 @@ let osc = oscillator (const Saw) (const 80.0)
 let note_clock = pulse ~frequency_hz:(const 4.0) ~duty_01:(const 0.5)
 
 let envelope =
-  asr_linear ~gate:note_clock ~attack_s:(const 0.2) ~release_s:(const 0.2)
+  ar_linear ~gate:note_clock ~attack_s:(const 0.2) ~release_s:(const 0.2)
 
 (* Apply some chebyshev filters to the oscillator. The [epsilon] argument
    controls how much resonance is present in the output. *)