+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## 0.0.2 (unreleased)
 
+- Rename ASR envelope generator to "AR"
 - Rename `Output_stream.create_with_downsample` to `Output_stream.create`
 
 ## 0.0.1